requirements.txt
numpy
scipy
librosa
matplotlib
requests
mir_eval
pandas
tqdm
scikit-learn
ipykernel
notebookpytest
